Extend error-checking for method lookup.
Previously the error-checking caught `AttributeError` in case there was
a failure to find a method name as key.  However, an error (the only
error?) that can happen is that the code finds all the matching methods,
finds them in a list comprehension (expected to be exactly one
element long), and then takes the first element of the resulting list.

If no match is found, the list is empty, and we get a `KeyError` trying
to take the first element of an empty list. I added a check for that
exception, as well.